Refactor neuropixel in a separate file and generate the library.
