+++ src/probeinterface/__init__.py
@@ -12,20 +12,22 @@ from .io import (
     write_prb,
     read_csv,
     write_csv,
-    read_imro,
-    write_imro,
     read_BIDS_probe,
     write_BIDS_probe,
     read_spikegadgets,
+    read_mearec,
+    read_nwb,
+    read_maxwell,
+    read_3brain,
+)
+from .neuropixel_tools import (
+    read_imro,
+    write_imro,
     read_spikeglx,
     parse_spikeglx_meta,
     get_saved_channel_indices_from_spikeglx_meta,
     read_openephys,
     get_saved_channel_indices_from_openephys_settings,
-    read_mearec,
-    read_nwb,
-    read_maxwell,
-    read_3brain,
 )
 from .utils import combine_probes
 from .generator import (
